Adding basic testing of the report handler
This would be a good time now for me to complain a little bit about the
complete lack of documentation and infrastructure around unit and integration
testing under the serverless framework. I've found a testing harness which
appears to be useful, but there is no clear way to me to be able to have
separate, clear unit tests on board without having to create more npm modules
and including them as more dependencies.